Final Update of Server-Side Lessons Unlimited

+++ LessonsUnlimited_V1.2/LessonsUnlimited_V1.2/Controllers/LessonsController.cs
@@ -3,6 +3,7 @@ using LessonsUnlimited_V1._2.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Web;
 using System.Web.Mvc;
 
@@ -10,11 +11,11 @@ namespace LessonsUnlimited_V1._2.Controllers
 {
     public class LessonsController : Controller
     {
+        
         private ILessonServices _service;
         public LessonsController(ILessonServices service)
         {
             _service = service;
-
         }
 
         // GET: Lessons
@@ -26,12 +27,22 @@ namespace LessonsUnlimited_V1._2.Controllers
         // GET: Lessons/Details/5
         public ActionResult Details(int id)
         {
-            return View();
+            return View(_service.Find(id));
         }
+
 
         // GET: Lessons/Create
         public ActionResult Create()
         {
+            // Employing claims to access user access to actions inside a contoller.
+            // Cast the user as a ClaimsIdentity
+
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit","true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
+            
             return View();
         }
 
@@ -39,20 +50,28 @@ namespace LessonsUnlimited_V1._2.Controllers
         [HttpPost]
         public ActionResult Create(Lesson lesson)
         {
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit", "true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
+
             if (ModelState.IsValid)
             {
                 _service.Create(lesson);
                 return RedirectToAction("Index");
-
             }
-            
                 return View();
-            
         }
 
         // GET: Lessons/Edit/5
         public ActionResult Edit(int id)
         {
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit", "true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
             var original = _service.Find(id);
             return View(original);
         }
@@ -61,6 +80,12 @@ namespace LessonsUnlimited_V1._2.Controllers
         [HttpPost]
         public ActionResult Edit(Lesson lesson)
         {
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit", "true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
+
             if (ModelState.IsValid)
             {
                 _service.Edit(lesson);
@@ -73,6 +98,12 @@ namespace LessonsUnlimited_V1._2.Controllers
         // GET: Lessons/Delete/5
         public ActionResult Delete(int id)
         {
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit", "true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
+
             var original = _service.Find(id);
             return View(original);
         }
@@ -82,6 +113,11 @@ namespace LessonsUnlimited_V1._2.Controllers
         [ActionName("Delete")]
         public ActionResult DeleteReally(int id)
         {
+            var user = User.Identity as ClaimsIdentity;
+            if (!user.HasClaim("CanEdit", "true"))
+            {
+                return new HttpUnauthorizedResult("Hey, Go Away!");
+            }
             _service.Delete(id);
             return RedirectToAction("Index");
         }
